==> 1_Obtencion_datos.ipynb <==
## Cargo las librerias
import numpy as np
import pandas as pd
from pandas.io.json import json_normalize
 

##Cargo las librerias de omdb
import omdb
import requests

##Cargo librerias de scraping web
import re 
from bs4 import BeautifulSoup
import requests

##Para archivos y directorios
import io
import os
import chardet

##Librerias de tiempo
import time
from datetime import datetime
import time

##Libreria del zip
from zipfile import ZipFile

#Con esto podemos echar un vistazo a código HTML directamente en la Notebook
from IPython.display import HTML, display

##Librerías utilizadas para el muestreo de género de las peliculas
from collections import Counter
from random import uniform
from random import random
from bisect import bisect

###Librerias para el conteno de palabras (subtitulos)
from sklearn.feature_extraction.text import CountVectorizer
import nltk as nl
from nltk import word_tokenize
from nltk.stem.porter import PorterStemmer
import string as st
from textblob import TextBlob

##Separar training y test
from sklearn.model_selection import train_test_split, KFold, cross_val_score, GridSearchCV
from sklearn import preprocessing
##Metricas
from sklearn.metrics import confusion_matrix, roc_auc_score, roc_curve, accuracy_score

##Modelos
from sklearn.linear_model import LogisticRegression, LogisticRegressionCV
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB

from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.model_selection import cross_val_score


##Graficos
% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

##Importo el data set de películas
# Cargar la data
df_movies = pd.read_csv('./tmdb_5000_movies.csv', sep = ',')

print('df_movies: ',df_movies.shape[0])
print(df_movies.info())


## Filtro las peliculas para que solo me traiga las que el idioma original sean en ingles
df_movies.original_language.value_counts()
df_movies = df_movies.loc[(df_movies.original_language == 'en') & (~df_movies.release_date.isnull()), :]
df_movies.shape

##Cambio las fechas a datetime
df_movies.release_date = pd.to_datetime(df_movies.release_date)
print('min fecha: ', min(df_movies.release_date))
print('max fecha: ',max(df_movies.release_date))
df_movies.head(5)

##creo el campo anio
df_movies['anio']=df_movies['release_date'].dt.year.astype(int)


##Cargo el key de la api de omdb como default
omdb.set_default('apikey', 'c0b0cfca')


hora_ini = datetime.now()

data_base = pd.DataFrame()
##Consulto por película
for i in range(df_movies.shape[0]):
    try:
        peliculas = omdb.get(title=df_movies.iloc[i,6], media_type = 'movie', year = df_movies.iloc[i, 20], fullplot=True, tomatoes=True, timeout=60)
        df = pd.DataFrame.from_dict(json_normalize(peliculas), orient='columns')
        df['id_movie'] = df_movies.iloc[i, 3]
        data_base = pd.concat([data_base, df])
        time.sleep(1)
        if i%100 == 0:
            print(i)
            print('Duration: {}'.format(datetime.now() - hora_ini))
    except Exception as e:
        print("Unexpected error raised")
        print(e, "\n")
        continue

df_movies.info()

#Borro los que no pude detectar
data_movies = data_base.loc[~(data_base.id_movie.notnull()&data_base.title.isnull()),:]

##Me fijo en las peliculas que no pudo detectar
faltantes = pd.merge(left=df_movies, right=data_movies, how= 'left',left_on='id', right_on='id_movie', indicator=True)
faltantes = faltantes.loc[faltantes['_merge'] == 'left_only', ['budget', 'genres', 'homepage', 'id', 'keywords', 'original_language',
       'original_title', 'overview', 'popularity', 'production_companies',
       'production_countries', 'release_date', 'revenue', 'runtime',
       'spoken_languages', 'status', 'tagline', 'title', 'vote_average',
       'vote_count', 'anio']]


faltantes.shape

##Recupero los faltantes
##Cargo el key de la api de omdb como default
omdb.set_default('apikey', 'c0b0cfca')


hora_ini = datetime.now()

data_base = pd.DataFrame()
##Consulto por película
for i in range(faltantes.shape[0]):
    try:
        peliculas = omdb.get(title=faltantes.iloc[i,6], media_type = 'movie', year = faltantes.iloc[i, 20], fullplot=True, tomatoes=True, timeout=60)
        df = pd.DataFrame.from_dict(json_normalize(peliculas), orient='columns')
        df['id_movie'] = faltantes.iloc[i, 3]
        data_base = pd.concat([data_base, df])
        time.sleep(1)
        if i%10 == 0:
            print(i)
            print('Duration: {}'.format(datetime.now() - hora_ini))
    except Exception as e:
        print("Unexpected error raised")
        print(e, "\n")
        continue

data_movies = pd.concat([data_movies, data_base.loc[data_base['title'].notnull(),:]])

data_movies.info()

##Guardo los imdb_id y los titulos
data_movies.loc[:,['title','imdb_id']].to_csv('imdb_id.csv', sep=';', index = False)

##Guardo las peliculas
data_movies.to_csv('data_movies.csv', sep=';', index = False)

data_movies.columns
data_movies.iloc[1,20]

##Importo el archivo
data_movies = pd.read_csv('./imdb_id2.txt', sep = ';')


data_movies.columns

###Traigo los subtitulos
hora_ini = datetime.now()
subtitulos_encontrados = pd.DataFrame()

head = {'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) '
           'Chrome/59.0.3071.115 Safari/537.36'}

for i in range(2,100):
    try:
        s = requests.Session()
        
        ##Pagina para buscar el id de opensubtitles y para bajar el subtitulo
        search_url = 'https://www.opensubtitles.org/en/search/sublanguageid-eng/imdbid-'
        download_url = 'https://www.opensubtitles.org/en/download/vrf-108d030f/sub/'

        ##Hago el requests get cuando le paso el id de imdb
        #contenido = requests.get(search_url + data_movies.iloc[i,8], headers = head, auth=('alquimista1978','alqu1m1sta'))
        contenido = s.get(search_url + data_movies.iloc[i,1], headers = head, auth=('alquimista1978','alqu1m1sta'))


        ##Hago BeautifulSoup para estructurar la paguina
        soup = BeautifulSoup(contenido.text, 'html.parser')

        ##Busco la clase bnone y extraigo el href donde se encuentra el id de opensubtitles
        id_interno = soup.find('a', attrs={'class': 'bnone'}).get('href')
        ##Busco del href los numeros
        id_interno = re.findall('\\d+', id_interno)[0]
        
        time.sleep(np.random.uniform(1,10))
       
        ##genero el link para bajar el subtitulo con el id interno de opensubtitles
        #download_url = requests.get(download_url + id_interno, headers = head, auth=('alquimista1978','alqu1m1sta'))
        download_url = s.get(download_url + id_interno, headers = head, auth=('alquimista1978','alqu1m1sta'))
        
        ##Abro el zip del subtitulo 
        z = zipfile.ZipFile(io.BytesIO(download_url.content)) # Opening zip file

        ##Dentro del zip encuentro el archivo con extension srt
        srt = [m for m in z.namelist() if re.findall(pattern='.srt', string = m)][0]
        
        ##Guardo un resumen de lo encontrado
        s_econt = pd.DataFrame({'imdb_id': data_movies.iloc[i,1], 'title': data_movies.iloc[i,0], 'subtitulo':srt}, index= {i})
    
        subtitulos_encontrados = pd.concat([subtitulos_encontrados, s_econt])
        
        ##Del zip extraigo solo el subtitulo y lo reescribo con el id de imdb
        for f in z.filelist:
            if f.filename == srt:
                f.filename = data_movies.iloc[i,8] +'.srt'
                z.extract(member = f, path='./Subtitulos/')
        
        time.sleep(np.random.uniform(10,20))
        
        ##Control de tiempo, cuantos va ejecutando y cuantos subtitulos encontro
        if i%10 == 0:
            print('Cantidad de registros procesados:', i)
            print('Tiempo de ejecucion: {}'.format(datetime.now() - hora_ini))
            print('Cantidad de subtitulos encontrados:', len(np.unique(subtitulos_encontrados.subtitulo)))
    except Exception as e:
        print("Unexpected error raised")
        print(e, "\n")
        print(download_url)
        time.sleep(np.random.uniform(10,20))
        continue    

#Vemos el contenido renderizado
display(HTML(download_url.text))

###Traigo los subtitulos de otra forma
hora_ini = datetime.now()
subtitulos_encontrados = pd.DataFrame()

head = {'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) '
           'Chrome/59.0.3071.115 Safari/537.36'}

for i in range(171, data_movies.shape[0]):
    try:
        s = requests.Session()
        
        ##Pagina para buscar el id de opensubtitles y para bajar el subtitulo
        search_url = 'https://www.opensubtitles.org/en/search/sublanguageid-eng/imdbid-'
        download_url = 'https://www.opensubtitles.org/en/download/vrf-108d030f/sub/'

        ##Hago el requests get cuando le paso el id de imdb
        #contenido = requests.get(search_url + data_movies.iloc[i,8], headers = head, auth=('alquimista1978','alqu1m1sta'))
        contenido = s.get(search_url + data_movies.iloc[i,8], headers = head, auth=('alquimista1978','alqu1m1sta'))


        ##Hago BeautifulSoup para estructurar la paguina
        soup = BeautifulSoup(contenido.text, 'html.parser')

        ##Busco la clase bnone y extraigo el href donde se encuentra el id de opensubtitles
        id_interno = soup.find('a', attrs={'class': 'bnone'}).get('href')
        ##Busco del href los numeros
        id_interno = re.findall('\\d+', id_interno)[0]
        
        time.sleep(np.random.uniform(1,2))
       
        ##Guardo un resumen de lo encontrado
        s_econt = pd.DataFrame({'imdb_id': data_movies.iloc[i,8], 'title': data_movies.iloc[i,20], 'download_url':download_url + id_interno}, index= {i})
    
        subtitulos_encontrados = pd.concat([subtitulos_encontrados, s_econt])
        
        ##Control de tiempo, cuantos va ejecutando y cuantos subtitulos encontro
        if i%10 == 0:
            print('Cantidad de registros procesados:', i)
            print('Tiempo de ejecucion: {}'.format(datetime.now() - hora_ini))
            
    except Exception as e:
        print("Unexpected error raised")
        print(e, "\n")
        print(download_url)
        continue    

subtitulos_encontrados.to_csv('download.csv', sep=';', index = False)
