==> 4_Trabajo_subtitulos_genero_rated.ipynb <==
## Cargo las librerias
import numpy as np
import pandas as pd
from pandas.io.json import json_normalize
 

##Cargo las librerias de omdb
import omdb
import requests

##Cargo librerias de scraping web
import re 
from bs4 import BeautifulSoup
import requests

##Para archivos y directorios
import io
import os
import chardet

##Librerias de tiempo
import time
from datetime import datetime
import time

##Libreria del zip
from zipfile import ZipFile

#Con esto podemos echar un vistazo a código HTML directamente en la Notebook
from IPython.display import HTML, display

##Librerías utilizadas para el muestreo de género de las peliculas
from collections import Counter
from random import uniform
from random import random
from bisect import bisect

###Librerias para el conteno de palabras (subtitulos)
from sklearn.feature_extraction.text import CountVectorizer
import nltk as nl
from nltk import word_tokenize
from nltk.stem.porter import PorterStemmer
import string as st
from textblob import TextBlob

##Separar training y test
from sklearn.model_selection import train_test_split, KFold, cross_val_score, GridSearchCV
from sklearn import preprocessing
##Metricas
from sklearn.metrics import confusion_matrix, roc_auc_score, roc_curve, accuracy_score

##Modelos
from sklearn.linear_model import LogisticRegression, LogisticRegressionCV
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB

from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.model_selection import cross_val_score


##Graficos
% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

#################################################################################################################################
##Importo los archivo
#################################################################################################################################
sub_text = pd.read_csv('./Subtitulos_limpios.csv', sep = ';')
#data_movies = pd.read_csv('./data_movies.csv', sep = ';')
data_movies = pd.read_csv('./data_movies_genero_V2.csv', sep = ';')



##########################################################################################################
########################################## Trabajo con el género #########################################
##########################################################################################################

##1) calculo la frecuencia de los generos individuales
#Parto cada uno de los generos en una listas de listas (una lista de generos por pelicula)
generos = [i.split(', ') for i in data_movies['genre']]
##Creo una sola lista de las listas por peliculas
generos_lista = sum(generos, [])
##Hay en total 10.830 registros

##Calculo las probabilidades de ocurrencia cada uno de los generos
generos = Counter(generos_lista)
generos = [(i, generos[i] / len(generos_lista)) for i in generos]

##Funcion que toma los generos con sus probabilidades y eligue uno 
def weighted_choice(choices):
    values, weights = zip(*choices)
    total = 0
    cum_weights = []
    for w in weights:
        total += w
        cum_weights.append(total)
    x = random() * total
    i = bisect(cum_weights, x)
    return values[i]

##A cada genero de las peliculas le pego su probabilidad
pos = [z[0] for z in generos]
list_generos=[]
for i in [i.split(', ') for i in data_movies['genre']]:
    list1 = []
    for j in i:
        list0 = (j, generos[pos.index(j)][1])
        list1.append(list0)
    list_generos.append(list1)

##Muestreo los generos y se los agrego al dataset
data_movies['genero'] = [weighted_choice(i) for i in list_generos]

###Compruebo que al menos el genero que muestreo este entre los generos posibles
genre = [i.split(', ') for i in data_movies['genre']]
genero = data_movies['genero']
sum([genero[i] in genre[i]  for i in range(len(genero))])
##Como la suma me da 4318 que es la cantidad total de registros del data set data_movie, esto quiere decir que esta correcto el genero muestreado

##################################################################################################################################
########## Arreglo el dataset data_movies
##################################################################################################################################
##Borro los campos que estan en NULL
#data_movies = data_movies.drop(['tomato_consensus', 'tomato_fresh','tomato_image','tomato_meter','tomato_rating','tomato_reviews','tomato_rotten','tomato_user_meter','tomato_user_rating','tomato_user_reviews'], axis=1)

##Reemplazo los N/A por nan
data_movies = data_movies.replace('N/A', np.nan)

##Transformo las fechas en date
data_movies.released = pd.to_datetime(data_movies.released)

## Elimino la palabra min a la Duracion
def intminutes(x):
    if pd.isnull(x):
        return np.nan
    x = str(x)
    y = x.replace('min', '').strip()
    return int(y)
data_movies.runtime = data_movies.runtime.apply(intminutes)

##Paso a entero el año
data_movies.year = data_movies.year.astype(int)

## Paso el imdbrating a float
data_movies.imdb_rating = data_movies.imdb_rating.astype(float)

##Le saco las comas a imdb_votes
def intvotes(x):
    if pd.isnull(x):
        return np.nan
    x = str(x)
    y = x.replace(',', '').strip()
    return int(y)
#data_movies.imdb_votes = data_movies.imdb_votes.apply(intvotes)

##Guardo el data set data_movie con el genero muestreado
#data_movies.to_csv('data_movies_genero_V2.csv', sep=';', index = False)

###############################################################################################################
########################################## Trabajo con los subtitulos #########################################
###############################################################################################################
##Analizo el idioma del texto y muestro solamente los que no son ingles
idioma=[]
for i in sub_text.subtitulo:
    b = TextBlob(i)
    idio = b.detect_language()
    idioma.append(idio)
    if not(idio == 'en'):


#sub_text.subtitulo[2679]
#data_movies.loc[data_movies.imdb_id == sub_text.imdb_id[2679],:]
#sub_text.imdb_id[724]        print(len(idioma)-1, idio)

##Borro los subtitulos que no estan en ingles
sub_text = sub_text.drop(sub_text.index[[21,187,216,318,366,452,468,597,689,724,882,944,1001,1190,1225,1245,1290,1407,1444,1500,1508,1568,1571,1670,1726,1730,1765,1882,1944,2190,2202,2203,2206,2308,2330,2371,2448,2549,2581,2674,2679,2700,3000,3064,3066]])

sub_text.head(50)[19:25]

##Reseteo el indice
sub_text = sub_text.reset_index(drop=True)

sub_text.head(50)[8:25]

##Limpio los subtitulos
for i in range(sub_text.shape[0]):
    numeros = re.compile('\d*')
    comillas2 = re.compile("''")
    comillas = re.compile("'")
    espacios = re.compile("\s{2,}")
    enter = re.compile("\n{2,}")
    solo_enter = re.compile ("\n")
    
    texto = sub_text.iloc[i, 1]
    texto = numeros.sub('', texto)
    texto = comillas2.sub('', texto)
    texto = comillas.sub('', texto)
    texto = espacios.sub(' ', texto)
    texto = enter.sub('\n', texto)
    texto = solo_enter.sub(' ', texto).strip()
    sub_text.iloc[i, 1] = texto

sub_text[sub_text.index==8]

sub_text.shape

hora_ini = datetime.now()

##Funciones para tokenizar y hacer stemmer
stemmer = PorterStemmer()
def stem_tokens(tokens, stemmer):
    stemmed = []
    for item in tokens:
        stemmed.append(stemmer.stem(item))
    return stemmed

def tokenize(text):
    tokens = nl.word_tokenize(text)
    tokens = [i for i in tokens if i not in st.punctuation]
    stems = stem_tokens(tokens, stemmer)
    return stems

##Hago countvectorizer con los subtitulos y creo un dataframe con los mismos
cv = CountVectorizer(tokenizer=tokenize, stop_words='english', lowercase=True, min_df=10, max_df=0.8)
#cv = CountVectorizer(stop_words='english', lowercase=True, min_df=10, max_df=0.90)

data = cv.fit_transform(sub_text.subtitulo).todense()

print('Duration: {}'.format(datetime.now() - hora_ini))


len(cv.get_feature_names())

#Genero un data frame con las palabras
columns = [c for c in cv.get_feature_names()]
subdf = pd.DataFrame(data, columns=columns)
subdf.shape

subdf.iloc[8,:].sum()

### Le pego el imdb_id
subdf['imdb_id'] = sub_text['imdb_id']

subdf.loc[subdf.index==8, 'music']

##Le pego los datos del target
movie_completo = pd.merge(left = subdf, right = data_movies, on='imdb_id')

movie_completo.shape

##Guardo el data set
movie_completo.to_csv('4_movie_completo.csv', sep=';', index = False)

movie_completo.shape
