==> 7_Recomendacion.ipynb <==
## Cargo las librerias
import numpy as np
import pandas as pd
from scipy.spatial.distance import squareform, pdist

##Librerias de tiempo
import time
from datetime import datetime
import time

import re

##Graficos
% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
from math import pi

#################################################################################################################################
##Importo el archivo
#################################################################################################################################
movie_sent = pd.read_csv('6_peliculas_sentimientos.csv', sep = ';')

movie_sent.shape

##Filtro para quedarme con los campos necesarios para armar la matriz de distancia
movie_filter=movie_sent[['title','fear', 'trust', 'joy', 'anticipation', 'anger', 'disgust', 'sadness','surprise']]

##Creo la matriz de distancia de todas las películas
mat_dist=pd.DataFrame(squareform(pdist(movie_filter.iloc[:, 1:], metric='euclidean')), columns=movie_filter.title, index=movie_filter.title)

##Busco las peliculas similares en sentimiento a una dada película
#pelicula_buscada = 'X-Men: Days of Future Past'
pelicula_buscada = 'Paranormal Activity'
#pelicula_buscada = 'Ratatouille'
peliculas_similares = pd.DataFrame(mat_dist[pelicula_buscada].sort_values( ).head(9)[1:])

pelicula_grafico = movie_sent.loc[movie_sent.title.isin(peliculas_similares.index.tolist() + [pelicula_buscada]), :]
peliculas_similares

##Grafico las películas similares
hora_ini = datetime.now()
cant=0

for i in pelicula_grafico['title']:
    
    ##Filtro la pelicula
    pel_filt = pelicula_grafico.loc[pelicula_grafico.title == i, ['fear', 'trust','joy', 'anticipation','anger', 'disgust','sadness', 'surprise']]
    ## Grafico
    
    sent = pel_filt.columns
    values = pel_filt.values.tolist()[0]

    N = len(sent)

    values += values[:1]
    values

    # What will be the angle of each axis in the plot? (we divide the plot / number of variable)
    angles = [n / float(N) * 2 * pi for n in range(N)]
    angles += angles[:1]

    # Initialise the spider plot
    ax = plt.subplot(111, polar=True)

    # Draw one axe per variable + add labels labels yet
    plt.xticks(angles[:-1], sent, color='black', size=12)

    # Draw ylabels
    ax.set_rlabel_position(90)
    plt.yticks([10,20, 30, 40], ["10","20", "30", "40"], color="grey", size=10)
    plt.ylim(0,50)

    # Plot data
    ax.plot(angles, values, linewidth=1, linestyle='solid')

    # Fill area
    ax.fill(angles, values, 'b', alpha=0.1)

    # Agrego el titulo
    ax.axes.set_title(i,fontsize=20)

    #Guardo el grafico
    plt.savefig('./graficos_sentimientos/Recomendacion/'+re.sub('\W+', ' ', i)+'.png')
    plt.close()
    
    cant += 1
    print('cant_procesadas:',cant, 'de',len(pelicula_grafico['title']) ,'Duration: {}'.format(datetime.now() - hora_ini))
